Refactor storage into item and delivery
