+++ src/main/java/seedu/address/MainApp.java
@@ -26,10 +26,10 @@ import seedu.address.model.inventorymodel.InventoryModel;
 import seedu.address.model.inventorymodel.InventoryModelManager;
 import seedu.address.model.inventorymodel.ReadOnlyInventoryBook;
 import seedu.address.model.util.SampleDataUtil;
-import seedu.address.storage.DeliveryBookStorage;
-import seedu.address.storage.InventoryBookStorage;
-import seedu.address.storage.JsonDeliveryBookStorage;
-import seedu.address.storage.JsonInventoryBookStorage;
+import seedu.address.storage.delivery.DeliveryBookStorage;
+import seedu.address.storage.item.InventoryBookStorage;
+import seedu.address.storage.delivery.JsonDeliveryBookStorage;
+import seedu.address.storage.item.JsonInventoryBookStorage;
 import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.Storage;
 import seedu.address.storage.StorageManager;
