+++ src/main/java/seedu/address/storage/Storage.java
@@ -9,6 +9,8 @@ import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.deliverymodel.ReadOnlyDeliveryBook;
 import seedu.address.model.inventorymodel.ReadOnlyInventoryBook;
+import seedu.address.storage.delivery.DeliveryBookStorage;
+import seedu.address.storage.item.InventoryBookStorage;
 
 /**
  * API of the Storage component
