+++ src/main/java/seedu/address/storage/delivery/DeliveryBookStorage.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.delivery;
 
 import java.io.IOException;
 import java.nio.file.Path;
