+++ src/main/java/seedu/address/storage/item/InventoryBookStorage.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.item;
 
 import java.io.IOException;
 import java.nio.file.Path;
