+++ src/main/java/seedu/address/storage/item/JsonAdaptedItem.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.item;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -8,7 +8,6 @@ import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.item.Item;
 import seedu.address.model.item.Metric;
