+++ src/main/java/seedu/address/storage/item/JsonInventoryBookStorage.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.item;
 
 import static java.util.Objects.requireNonNull;
