+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -122,7 +122,8 @@ public class MainWindow extends UiPart<Stage> {
         resultDisplay = new ResultDisplay();
         resultDisplayPlaceholder.getChildren().add(resultDisplay.getRoot());
 
-        StatusBarFooter statusBarFooter = new StatusBarFooter(logic.getInventoryBookFilePath());
+        StatusBarFooter statusBarFooter = new StatusBarFooter(logic.getInventoryBookFilePath(),
+                logic.getDeliveryBookFilePath());
         statusbarPlaceholder.getChildren().add(statusBarFooter.getRoot());
 
         CommandBox commandBox = new CommandBox(this::executeCommand);
